requirements.txt
aiopg==1.4.0
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-redis==5.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
fastapi==0.88.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.5.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
lxml==4.9.2
outcome==1.2.0
packaging==22.0
peewee==3.15.4
peewee-async==0.8.0
priority==2.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
py==1.11.0
pycparser==2.21
pydantic==1.10.4
pyshark==0.5.3
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
redis==4.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
starlette==0.22.0
toml==0.10.2
trio==0.22.0
typing_extensions==4.4.0
tzdata==2022.7
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
wsproto==1.2.0
